requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
bg-run==0.0.1
bleach==2.0.0
bs4==0.0.1
click==6.7
-e git+https://jjmontesl@github.com/jjmontesl/cubes.git@151256cb2b41d8b36b2ae080afd2a5c67c63e1a2#egg=cubes
-e git+https://jjmontesl@github.com/jjmontesl/cubesextensions.git@6926cac3752dafde42b2326d9dbc1048b7c47a9c#egg=cubes_extensions
-e git+https://jjmontesl@github.com/jjmontesl/cubetl.git@41bce29ce31d23f328f34f5e966fef6d860065f7#egg=CubETL
decorator==4.0.11
entrypoints==0.2.2
enum34==1.1.6
expressions==0.2.3
Flask==0.12.1
fs==2.0.4a1
future==0.16.0
grako==3.22.0
html5lib==0.999999999
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lxml==3.7.3
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.12.1
pandas==0.19.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyreadline==2.1
python-dateutil==2.6.0
python-slugify==1.2.4
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
repoze.lru==0.6
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.2.12
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
Unidecode==0.4.20
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.1
widgetsnbextension==2.0.0
